% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCapstone.R
\docType{data}
\name{data_raw}
\alias{data_raw}
\title{Reading the NOAA earthquake data file}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 1 columns.}
\usage{
data_raw
}
\description{
Reading the NOAA earthquake data file
}
\keyword{datasets}
