% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleans raw data from U.S. National Oceanographic and Atmospheric Administation's (NOAA)}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{A data frame with raw data obtained from NOAA website}
}
\value{
A data.frame containing the cleaned NOAA data
}
\description{
This function takes data from NOAA website as an input
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}. It adds
a column DATE with cleaned date, transforms LATITUDE and
LONGITUDE columns as numeric objects.
}
\examples{
\dontrun{
data <- readr::read_delim("signif.txt", delim = "\\t")
clean_data <- eq_clean_data(data)
}

}
