% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates a label for a map of earthquake}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A data frame containing cleaned NOAA data}
}
\value{
A character vector with labels
}
\description{
This function creates a label for the map based on location
name, magnitude and casualties from NOAA earthquake data
}
\examples{
\dontrun{
eq_create_label(clean_data)
}

}
