% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_data_read}
\alias{eq_data_read}
\title{Mastering Software Development in R Specialization Capstone Project
Coursera Capstone Project
The overall goal of the capstone project is to integrate the skills you have developed
over the courses in this Specialization and to build a software package that can be used
to work with the NOAA Significant Earthquakes dataset.
Eason Shen
redcomet97@outlook.com
Documentation at https://github.com/jeeweshjha/Mastering-Software-Development-in-R-Capstone
Use the Use_vignette once to generate the vignette folder in the package's skeleton where the vignette documentation
made with rmarkdown will be made
use_vignette("capstoneproject")
Use the use_testthat() once to generate the test folder in the package's skeleton where the test scripts will be placed}
\usage{
eq_data_read(filename)
}
\arguments{
\item{filename}{The name of the NOAA earthquake data file}
}
\value{
This function returns tbl_df object (earthquake data)
}
\description{
use_testthat()
}
\details{
Function for reading the NOAA earthquake data file
}
\note{
The function will stop If the filename does not exist (error message)
}
\examples{
\dontrun{
filename<-system.file("data","earthquakes_data.txt.zip",package="capstone")
eq_data_read(filename)
}

}
