% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module1.R
\name{eq_date_clean}
\alias{eq_date_clean}
\title{Creates a Date variable from year, month, day where year could be positive (AD) or negative (BC)}
\usage{
eq_date_clean(input_year, input_month, input_day)
}
\arguments{
\item{input_year}{An integer (positive or negative) indicating the year. A negative integer represents BC}

\item{input_month}{An integer indicating the month. If NA, it defaults to 1}

\item{input_day}{An integer indicating the day of the month. If NA, it defaults to 1}
}
\value{
This function returns a Date variable
}
\description{
Creates a Date variable from year, month, day where year could be positive (AD) or negative (BC)
}
\examples{
\dontrun{
eq_date_clean(2015, 3, 30)
eq_date_clean(-2249, 1, 31)
}

}
