% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Function to clean the LOCATION_NAME column in the NOAA
earthquake dataset}
\usage{
eq_location_clean(raw_data)
}
\arguments{
\item{raw_data}{A data frame with raw data obtained from NOAA website}
}
\value{
A data frame with cleaned LOCATION_NAME column
}
\description{
Function to clean the LOCATION_NAME column in the NOAA
earthquake dataset
}
\examples{
\dontrun{
raw_data <- readr::read_delim("signif.txt", delim = "\\t")
clean_data <- eq_location_clean(raw_data)
}

}
