% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Creates a map earthquake}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A data frame containing cleaned NOAA earthquake data}

\item{annot_col}{A character. The name of the column in the data}
}
\value{
generates a map of earthquakes, The size of the circles reflects the magnitude of the earthquake.
}
\description{
This function creates a leaflet map of selected earthquakes based on
NOAA earthquake cleaned data.
}
\examples{
\dontrun{
eq_map(clean_data, annot_col = "LOCATION_NAME")
}

}
