% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_Geoms.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline function}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
This function is used in addition to GeomTimeline
See the layout contained within the course materials p.g. 452
Here the layout of this function is very similar.
This geom looks to chart a timeline of earthquakes for a given country / countries
with points (in the example) representing earthquake events, point size indicating
earthquake magnitude and colour representing number of deaths. x (the date) is a
required aesthetic whereas y (country) is optional.
}
\examples{
\dontrun{
NOAA_df_China_USA <- NOAA_df \%>\%
 filter(COUNTRY \%in\% c("CHINA", "USA"), YEAR > 2000)
ggplot(NOAA_df_China_USA, aes(x = DATE, y = COUNTRY,
                              color = as.numeric(TOTAL_DEATHS),
                              size = as.numeric(EQ_PRIMARY),
                              label = CLEAN_LOCATION_NAME)) +
  geom_timeline() +
  labs(size = "Richter scale value", color = "# deaths") +
  ggplot2::theme(panel.background = ggplot2::element_blank(),
                 legend.position = "bottom",
                 axis.title.y = ggplot2::element_blank()) +
  ggplot2::xlab("DATE")
}
}
