% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_Geoms.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label function}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  n_max = 5,
  ...
)
}
\arguments{
\item{n_max}{}
}
\description{
This function is used to label the points in the graph
}
\examples{
\dontrun{
NOAA_df_China_USA <- NOAA_df \%>\%
filter(COUNTRY \%in\% c("CHINA", "USA"), YEAR > 2000)
ggplot(NOAA_df_China_USA , aes(x = DATE, y = COUNTRY,
                               color = as.numeric(TOTAL_DEATHS),
                               size = as.numeric(EQ_PRIMARY),
                               label = CLEAN_LOCATION_NAME)) +
  geom_timeline() +
  labs(size = "Richter scale value", color = "# deaths") +
  ggplot2::theme(panel.background = ggplot2::element_blank(),
                 legend.position = "bottom",
                 axis.title.y = ggplot2::element_blank()) +
  ggplot2::xlab("DATE") +
  geom_timeline_label(data=NOAA_df_China_USA)
}
}
