% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_Geoms.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline class}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimeline
}
\description{
building a GeomTimeline class which is used in a function
}
\keyword{datasets}
