% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_Geoms.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{GeomTimelineLabel class}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimelineLabel
}
\description{
This is a GeomTimelineLabel class which is used in the
functions
}
\examples{
\dontrun{
NOAA_df_China_USA <- NOAA_df \%>\%
filter(COUNTRY \%in\% c("CHINA", "USA"), YEAR > 2000)
ggplot(NOAA_df_China_USA , aes(x = DATE, y = COUNTRY,
                               color = as.numeric(TOTAL_DEATHS),
                               size = as.numeric(EQ_PRIMARY),
                               label = CLEAN_LOCATION_NAME)) +
  geom_timeline() +
  labs(size = "Richter scale value", color = "# deaths") +
  ggplot2::theme(panel.background = ggplot2::element_blank(),
                 legend.position = "bottom",
                 axis.title.y = ggplot2::element_blank()) +
  ggplot2::xlab("DATE") +
  geom_timeline_label(data=NOAA_df_China_USA)
}
}
\keyword{datasets}
