% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/good_date.R
\name{good_date}
\alias{good_date}
\title{Create good Date}
\usage{
good_date(year, month, day)
}
\arguments{
\item{year}{vector of years (integer or cohercible to an integer
vector)}

\item{month}{vector of months (integer or cohercible to an integer
vector)}

\item{day}{vector of days (integer or cohercible to an integer
vector)}
}
\value{
a vector of Date
}
\description{
Create good Date
}
\examples{
good_date(1981, 04, 13)
good_date(
    c(1981, 1982),
    c(4, 5),
    c(13, 14)
)

good_date(
    c(1981, -1981),
    c(4, 5),
    c(13, 14)
)
}
