% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_Clean_data.R
\name{load_data}
\alias{load_data}
\title{Loads raw data}
\usage{
load_data(path)
}
\value{
Returns a dataframe result after calling read_delim
}
\description{
Loads raw data from a file folder
}
\examples{
\dontrun{
  NOAA_df_raw<-load_data()
}


}
