% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noaa_data}
\alias{noaa_data}
\title{NOAA Significant Earthquake Database}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6004 rows and 45 columns.}
\usage{
noaa_data
}
\description{
A dataset containing information on destructive earthquakes from 2150 BC to
the present.
}
\references{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\keyword{datasets}
