% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noaa.data}
\alias{noaa.data}
\title{Significant earthquakes data.}
\format{
A data frame with 6205 rows and 47 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
\url{http://dx.doi.org/10.7289/V5TD9V7K}
}
\usage{
noaa.data
}
\description{
A dataset contains information on destructive earthquakes from 2150 B.C. to the present that meet at least one of the following criteria: Moderate damage (approximately $1 million or more), 10 or more deaths, Magnitude 7.5 or greater, Modified Mercalli Intensity X or greater, or the earthquake generated a tsunami.
}
\examples{
\dontrun {
noaa.data
}
}
\keyword{datasets}
