% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOAA}
\alias{NOAA}
\title{NOAA dataset of all large earthquakes in recorded history.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5960 rows and 47 columns.}
\source{
see the readme file.
}
\usage{
NOAA
}
\description{
NOAA dataset of all large earthquakes in recorded history.
}
\examples{
\dontrun{
NOAA
}
}
\keyword{datasets}
