% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_Geoms.R
\name{plot_earthquakes_timeline}
\alias{plot_earthquakes_timeline}
\title{This is used to plot the earthquake timeline}
\usage{
plot_earthquakes_timeline(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
This is used to plot the earthquake timeline
}
\examples{
\dontrun{
NOAA_df_China_USA <- NOAA_df \%>\%
filter(COUNTRY \%in\% c("CHINA", "USA"), YEAR > 2000)
plot_earthquakes_timeline(NOAA_df_China_USA)
}

}
