% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_Geoms.R
\name{plot_earthquakes_timeline_label}
\alias{plot_earthquakes_timeline_label}
\title{plotting the earthquake timeline}
\usage{
plot_earthquakes_timeline_label(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
plotting the earthquake timeline using a dataframe
}
\examples{
\dontrun{
NOAA_df_China_USA <- NOAA_df \%>\%
filter(COUNTRY \%in\% c("CHINA", "USA"), YEAR > 2000)
plot_earthquakes_timeline_label(NOAA_df_China_USA)
}

}
