% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_data}
\alias{raw_data}
\title{raw_data}
\format{
A data frame with columns:
\describe{
 \item{YEAR}{Year.}
 \item{MONTH}{Month.}
 \item{DAY}{Day.}
 \item{HOUR}{Hour.}
 \item{MINUTE}{Minute.}
 \item{SECOND}{Second.}
 \item{COUNTRY}{Country.}
 \item{EQ_MAG_MW}{Magnitude.}
 \item{DEATHS}{Deaths.}
 \item{LOCATION_NAME}{Location name.}
 \item{INTENSITY}{Intensity.}
 \item{LATITUDE}{Latitude.}
 \item{LONGITUDE}{Longitude.}
 \item{EQ_PRIMARY}{Eq primary.}
 \item{TOTAL_DEATHS}{Total deaths.}
}
}
\source{
NOAA
}
\usage{
raw_data
}
\description{
Earthquakes raw data
}
\examples{
\dontrun{
 raw_data
}
}
\keyword{datasets}
