% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Creates a theme for timeline visualization}
\usage{
theme_timeline()
}
\description{
Creates a theme for timeline visualization
}
\examples{
\dontrun{
data \%>\% eq_clean_data() \%>\%
filter(COUNTRY \%in\% c("USA", "CHINA")) \%>\%
   ggplot(aes(x = DATE,
              y = COUNTRY,
              color = as.numeric(TOTAL_DEATHS),
              size = as.numeric(EQ_PRIMARY)
   )) +
   geom_timeline() +
   theme_timeline()
}

}
