% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Imports the U.S. National Oceanographic and Atmospheric Administation's (NOAA) 
dataset of significant earthquakes. For more info, see the readme file.}
\usage{
eq_clean_data(filename = NOAA)
}
\arguments{
\item{filename}{represents the dataset; defaults to its current directory.}
}
\description{
The package imports the data, cleans it, and then produces visualizations of earthquake
 data in select countries beginning in 2000.
}
\details{
\code{eq_clean_data} reads in the data, cleans it, and reformats the date 
 variables
}
\examples{
eq_clean_data()

}
