% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCapstone.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Function for cleaning the NOAA earthquake data file}
\usage{
eq_clean_data(x)
}
\arguments{
\item{x}{is the dataframe of NOAA earthquake data}
}
\value{
a dataframe which contains the Eathquake data with a date column created by uniting the year, month, day
and converting it to the Date class, furthermore LATITUDE, LONGITUDE and EQ_PRIMARY columns are converted to numeric class
}
\description{
Function for cleaning the NOAA earthquake data file
}
\examples{
\dontrun{
eq_clean_data(data_loc_clean)
}
}
