% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean the earthquake data}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{A dataframe containing the information about historical earthquakes}
}
\value{
A dataframe with clean data about earthquakes
}
\description{
This function cleans the earthquakes dataset
}
\examples{
\dontrun{
eq_clean_data(raw_data)
}

}
