% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq-clean-data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean raw NOAA data}
\usage{
eq_clean_data(path)
}
\arguments{
\item{path}{The path to the raw \code{signif.txt} file as string}
}
\value{
A tibble containing a cleaned data frame of the raw file
}
\description{
Clean raw NOAA data
}
\examples{
noaa_data <- eq_clean_data("data-raw/signif.txt")
}
