% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_Clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{cleans the loaded data}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{datafram of raw NOAA dataset}
}
\value{
Returns cleaned dataframe
}
\description{
It takes the raw data as input and then does the following things: 1.
1. Creates a date column using year month and day
2. Converts Latitude and longitude columns into numeric
3. Converts FLAG_TSUNAMI column
4. Calls eq_location_clean to clean the location column
}
\examples{
\dontrun{
NOAA_df_raw<-load_data()
NOAA_df<-eq_clean_data(NOAA_df_raw)
}


}
