% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Parameters and libraries needed for reading and cleaning the Earthquake data function}
\usage{
eq_clean_data(datafram)
}
\arguments{
\item{datfram}{is the dataframe that contains location names written in Uper case}
}
\value{
a dataframe which contains the Eathquake data filtered required for mapping in a timeline the data
}
\description{
Parameters and libraries needed for reading and cleaning the Earthquake data function
}
\examples{
\dontrun{
filename<-system.file("data","earthquakes_data.txt.zip",package="capstone")
eq_clean_data(eq_data_read(filename))
}

}
