% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module1.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Wrapper function to clean the NOAA data, specifically uniting the YEAR, MONTH and DAY columns, converting the LATITUDE and LONGITUDE columns from character to double and stripping out the country name from the LOCATION_NAME column}
\usage{
eq_clean_data(raw_noaa_df)
}
\arguments{
\item{raw_noaa_df}{Dataframe representing the raw NOAA dataset}
}
\value{
This function returns a dataframe with an additional column called "date" that combines YEAR, MONTH and DAY columns
}
\description{
Wrapper function to clean the NOAA data, specifically uniting the YEAR, MONTH and DAY columns, converting the LATITUDE and LONGITUDE columns from character to double and stripping out the country name from the LOCATION_NAME column
}
\examples{
eq_clean_data(noaa.data)

}
