% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creating the "popup_text" parameter. Not meant for use.}
\usage{
eq_create_label()
}
\description{
\code{eq_create_label} is used by \code{eq_map} to create the popup_text parameter,
which allows \code{eq_map} to display all relevant data, rather than one statistic.
}
\examples{
eq_map('popup_text')
}
