% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Annotate Earthquake data}
\usage{
eq_create_label(x)
}
\arguments{
\item{x}{Data set containing the information}
}
\value{
This function returns the dataset with a new column.
}
\description{
This is a function to create a new annotated information label based on Location,
Size of Earthquake, and Number of Deaths.
If there is no information in those fields, the tag will not be displayed.
}
\examples{
library(dplyr)
library(lubridate)
raw_data<-system.file("extdata", "signif.txt", package = "RCapstone")
eq_clean_data(raw_data) \%>\% eq_location_clean() \%>\%
 dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
 dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
 eq_map(annot_col = "popup_text")

}
