% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates popup text for markers.}
\usage{
eq_create_label(eq_clean = NULL)
}
\arguments{
\item{eq_clean}{The clean earthquake data in a tbl_df object.}
}
\value{
This function returns a character vector containing popup text to be used in a leaflet visualization.
}
\description{
This function takes the dataset as an argument and creates an HTML label that can be used as the annotation text
in the leaflet map. This function should put together a character string for each earthquake that will show the
cleaned location (as cleaned by the eq_location_clean() function), the magnitude (EQ_PRIMARY), and the
total number of deaths (TOTAL_DEATHS), with boldface labels for each ("Location", "Total deaths", and "Magnitude").
}
\examples{
\dontrun{
data_clean \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")
}

}
