% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Pop-up annotations in interactive map}
\usage{
eq_create_label(clean_data)
}
\arguments{
\item{clean_data}{The clean earthquake data}
}
\value{
A character vector containing the pop-up text to be used in a leaflet visualization.
}
\description{
This function should put together a character string for each earthquake that will show the
cleaned location, the magnitude, and the total number of deaths (TOTAL_DEATHS),
with boldface labels for each ("Location", "Total deaths", and "Magnitude").
}
\examples{
\dontrun{
readr::read_delim("earthquakes.tsv.gz", delim = "\t") \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
}

}
