% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq-create-label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create pop-up label for map}
\usage{
eq_create_label(x)
}
\arguments{
\item{x}{A data.frame that contains earthquake data by NOAA and was cleaned
with eq_clean_data()}
}
\value{
A character vector holding the contents of a pop-up for each
  earthquake formatted as HTML
}
\description{
Create pop-up label for map
}
\examples{
labels <- eq_create_label(noaa_data)
}
