% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapping.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create earthquake labels}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
Returns a character vector
}
\description{
This function is used tp create earthquake labels
}
\examples{
\dontrun{
NOAA_df<-eq_clean_data(loaded_data)
df <- NOAA_df \%>\% filter(COUNTRY \%in\% c("MEXICO"), YEAR > 2000)
map2<-df \%>\% dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
print(map2)
}

}
