% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module3.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates a html label for leaflet map based on location name, magnitude and casualties from NOAA earthquake data}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A data frame containing cleaned NOAA earthquake data}
}
\value{
A character vector with labels
}
\description{
Creates a html label for leaflet map based on location name, magnitude and casualties from NOAA earthquake data
}
\details{
The input \code{data.frame} needs to include columns LOCATION_NAME, EQ_PRIMARY and TOTAL_DEATHS with the earthquake location, magintude and total casualties respectively.
}
\examples{
\dontrun{
eq_create_label(data)
}
}
