% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean country and city names}
\usage{
eq_location_clean(filename = "data/signif.txt")
}
\arguments{
\item{filename}{represents the dataset; defaults to its current directory.}
}
\description{
\code{eq_location_clean} Returns country and city names along with the I_D from
the original dataset.
}
\examples{
eq_location_clean('data/signif.txt') 

}
\seealso{
\code{\link{eq_clean_data}}
}
