% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Earthquake Location}
\usage{
eq_location_clean(x)
}
\arguments{
\item{x}{Earthquake dataset}
}
\value{
This function returns the modified dataset
}
\description{
This is a simple function that cleans the Earthquake dataset.
It removes country from the Location Name variable.
If there is no location then, then just Country is supplied.
}
\examples{
library(magrittr)
library(readr)
raw_data<-readr::read_tsv(system.file("extdata", "signif.txt", package = "RCapstone"))
eq_clean_data(raw_data) \%>\% eq_location_clean()

}
