% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCapstone.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Earthquake Location Field}
\usage{
eq_location_clean(x)
}
\arguments{
\item{x}{A data frame of NOAA significant earthquake data.}
}
\value{
A dataframe with the same supplied data, but with the
  LOCATION_NAME variable cleaned up to remove the country name that is
  supplied in the COUNTRY variable and convert
  the remainder of the words in LOCATION_NAME to Title Case.
}
\description{
The function takes a data frame of NOAA earthquake data and cleans
the LOCATION_NAME field. It removes the country name (of form
"COUNTRY:" from the LOCATION_NAME data, and converts the remainder of the field
to Title Case. This will make the location name easier to read when plotting
and mapping.
}
\examples{
\dontrun{
data_loc_clean <- eq_location_clean(data_raw)
}
}
