% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean the LOCATION_NAME column}
\usage{
eq_location_clean(raw_data)
}
\arguments{
\item{raw_data}{A dataframe containing the information about historical earthquakes}
}
\value{
A dataframe with a clean LOCATION_NAME column
}
\description{
This function cleans the LOCATION_NAME column
}
\examples{
\dontrun{
eq_location_clean(eq_clean_data(raw_data))
}

}
