% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq-location-clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean raw NOAA column LOCATION_NAME}
\usage{
eq_location_clean(col)
}
\arguments{
\item{col}{The column of an uncleaned \code{noaa_data} dataset that contains the locations
of earthquake erruptions}
}
\value{
A character vector containing a cleaned versioned of the locations
}
\description{
The locations of earthquake erruptions stored in a \code{noaa_data} dataset
are separated by colons. This function extracts the third field containing
the name of the city. The function is only used internally for \code{eq_clean_data()}.
}
