% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq-location-clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean raw NOAA column LOCATION_NAME}
\usage{
eq_location_clean(col)
}
\description{
Given the LOCATION_NAME column from a raw NOAA database, this function
removes the name of the country at the beginning of each entry.
}
