% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_Clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans the LOCATION_NAME column}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
Returns a cleaned dataframe with modified Location Name column
}
\description{
Cleans the location name column and strips out unwanted
characters
}
\examples{
\dontrun{
  df_location <- load_data() \%>\% eq_location_clean()
}

}
