% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module1.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Strips out country name (including colon) and concerts names to title case}
\usage{
eq_location_clean(input_location)
}
\arguments{
\item{input_location}{String denoting the raw location field in the NOAA dataset}
}
\value{
String with the country name stripped out and in title case
}
\description{
Strips out country name (including colon) and concerts names to title case
}
\examples{
\dontrun{
eq_location_clean("JORDAN:  BAB-A-DARAA,AL-KARAK")
}

}
