% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{eq_map}
\alias{eq_map}
\title{Mexico earthquake map}
\usage{
eq_map(annot_col = "popup_text", data = NOAA)
}
\value{
\code{eq_map} generates a map of earthquakes in Mexico since 2000;
the size of the circles reflects the magnitude of the earthquake.
}
\description{
Mexico earthquake map
}
\examples{
eq_map('Date')
}
