% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Earthquake Map}
\usage{
eq_map(x, annot_col)
}
\arguments{
\item{x}{Data set to display}

\item{annot_col}{Column that contained the annotated text that pops up.}
}
\value{
This function returns a leaflet to display
}
\description{
This is a method to map the earthquake data. It passes the data to leaflet
}
\examples{
library(dplyr)
library(lubridate)
raw_data<-system.file("extdata", "signif.txt", package = "RCapstone")
eq_clean_data(raw_data) \%>\% eq_location_clean() \%>\%
 dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
 eq_map(annot_col = "DATE")

}
