% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_map}
\alias{eq_map}
\title{Interactive map with earthquakes data}
\usage{
eq_map(clean_data, annot_col = "DATE")
}
\arguments{
\item{clean_data}{The clean earthquake data.}

\item{annot_col}{Column in clean_data with the annotations}
}
\value{
An interactive map.
}
\description{
The function maps the epicenters (LATITUDE/LONGITUDE) and annotates each point with in
pop up window containing annotation data stored in a column of the data frame.
}
\examples{
\dontrun{
readr::read_delim("earthquakes.tsv.gz", delim = "\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
eq_map(annot_col = "DATE")
}

}
