% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq-map.R
\name{eq_map}
\alias{eq_map}
\title{Create map with earthquakes}
\usage{
eq_map(x, annot_col)
}
\arguments{
\item{x}{A data.frame that contains earthquake data by NOAA and was cleaned '
with eq_clean_data() ' @param annot_col}
}
\value{
A leaflet map created by leaflet::leaflet that indicates each
  earthquake with a circle. A pop-up displays the content of \code{annot_col}.
}
\description{
Create map with earthquakes
}
