% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapping.R
\name{eq_map}
\alias{eq_map}
\title{Plot a leaflet map}
\usage{
eq_map(df, annot_col)
}
\arguments{
\item{df}{a dataframe}

\item{annot_col}{annotation column}
}
\value{
an interactive leaflet map
}
\description{
We take a subset of our earthquake data and plot this on a map
using leaflet. See the leaflet docs (?leaflet) for a few relevant examples.
The input dataframe requires columns including LONGITUDE, LATITUDE, EQ_PRIMARY.
The annotation column ('annot_col') is also a required argument (no defaults).
}
\examples{
\dontrun{
NOAA_df<-eq_clean_data(loaded_data)
df <- NOAA_df \%>\% filter(COUNTRY \%in\% c("MEXICO"), YEAR > 2000)
map1<-eq_map(df,annot_col = "date")
print(map1)
}

}
