% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_map}
\alias{eq_map}
\title{Earthquakes Data in an Interactive Map.}
\usage{
eq_map(eq_clean = NULL, annot_col = "datetime")
}
\arguments{
\item{eq_clean}{The clean earthquake data in a tbl_df object.}

\item{annot_col}{Column in the tbl_df object to be used for annotation.}
}
\value{
This function returns an interactive map.
}
\description{
The Earthquakes will be mapped centered with their latitude and
longitude "epicenter". The epicenter is annotated based on an annot_col which the user can specify.
In addition, if the user specifies "popup_text" then a call to eq_create_label generates
the appropriate text for the popup.
}
\note{
If an invalid column name is provided, the function provides a warning
and uses the LOCATION_NAME column as teh annotation column.
}
\examples{
\dontrun{
filename<-system.file("data","earthquakes_data.txt.zip",package="capstone")
eq_location_clean(eq_clean_data(eq_data_read(filename))) \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(datetime) >= 1980) \%>\%
eq_map(annot_col = "datetime")
}

}
\references{
\url{http://rstudio.github.io/leaflet/}
}
