% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module3.R
\name{eq_map}
\alias{eq_map}
\title{Creates a geographical map using OpenStreetMap leaflet tiles to display earthquake locations based on cleaned NOAA dataset}
\usage{
eq_map(data, annot_col = NULL)
}
\arguments{
\item{data}{A preproccessed tibble (data.frame) of NOAA earthquakes dataset}

\item{annot_col}{A character. The name of the column in the data that should be used as descriptor.}
}
\value{
A leaflet map with earthquakes and annotations.
}
\description{
Creates a geographical map using OpenStreetMap leaflet tiles to display earthquake locations based on cleaned NOAA dataset
}
\examples{
\dontrun{
eq_map(data, annot_col = "LOCATION_NAME")
}
}
