% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Graph the timeline}
\usage{
geom_timeline(data = NOAA)
}
\description{
\code{geom_timeline} returns a graph of earthquakes from 2000-2017 in the USA 
and China. Size corresponds to the Richter Magnitude of the earthquake
}
\examples{
geom_timeline()

}
