% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomTimeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Plot Timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{x}{Date of earthquakes}

\item{size}{[optional] Size of earthquake}

\item{color}{[optional] Number of Deaths}

\item{group}{[optional] Column indicating how the data could be grouped.}
}
\value{
This function returns a graphic object
}
\description{
This is a method of creating a graph that maps earthquake data along the timeline.
}
\examples{
library(dplyr)
library(ggplot2)
raw_data<-readr::read_tsv(system.file("extdata", "signif.txt", package = "RCapstone"))
eq_clean_data(raw_data) \%>\% eq_location_clean() \%>\%
dplyr::filter(!is.na(EQ_PRIMARY), !is.na(DEATHS)) \%>\%
ggplot2::ggplot() +
 ggplot2::aes(
   x = DATE,
   size = EQ_PRIMARY,
   colour = DEATHS
 ) +
 geom_timeline()

}
