% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Mapping}

\item{data}{Data}

\item{stat}{Stat}

\item{position}{Position}

\item{na.rm}{na.rm}

\item{show.legend}{Show legend}

\item{inherit.aes}{Inherit aes}

\item{...}{other arguments}
}
\value{
A timeline with all earthquakes between a set of dates for a set of countries.
}
\description{
This function uses GeomTimeline Prototype for plotting a timeline with the earthquakes of a given country
}
\examples{
\dontrun{
eq_location_clean(eq_clean_data(raw_data)) \%>\%
dplyr::filter(datetime >= "1960-01-01" & datetime <="2011-01-01" & COUNTRY == c("CHILE","USA"))\%>\%
ggplot() +
geom_timeline(aes(x = datetime, size = EQ_MAG_ML, colour = DEATHS, fill = DEATHS))
}

}
