% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module2.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Creates a new geom to graphically depict of timelines of earthquake events}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  xmin = NULL,
  xmax = NULL,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{data}{A preproccessed tibble (data.frame) of NOAA earthquakes}

\item{xmin}{a character vector indicating the mininum date specified for earthquakes time line plotting. Should follow the format of YYYY-MM-DD.}

\item{xmax}{a character vector indicating the maximum date specified for earthquakes time line plotting. Should follow the format of YYYY-MM-DD.}
}
\description{
Creates a new geom to graphically depict of timelines of earthquake events
}
\examples{
\dontrun{
Capstone::noaa.data \%>\%
   Capstone::eq_clean_data() \%>\%
   dplyr::filter(COUNTRY \%in\% c("MEXICO", "INDONESIA", "CHINA")) \%>\%
   ggplot(aes(x = date, y = COUNTRY, label = LOCATION_NAME,group = COUNTRY, size = EQ_MAG_MW, fill = DEATHS/1000)) +
   geom_timeline(xmin = xmindate) +
   geom_timeline_label(n_max = 10, xmin = xmindate) +
   theme_classic()+ labs(y = "") +
   theme(legend.position = "bottom", axis.line.y = element_blank(), axis.ticks.y= element_blank())
}

}
