% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{geom_timeline_label}
\alias{geom_timeline_label}
\alias{GeomTimelineLabel}
\title{Time line labels of earthquakes}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  n_max = NULL,
  ...,
  na.rm = FALSE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function. Cannot be jointy specified with
\code{nudge_x} or \code{nudge_y}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{n_max}{(int) number of earthquakes, where we take the n_max
largest (in \code{size}, if provided) earthquakes}

\item{...}{further arguments passed to the geom layer}

\item{na.rm}{(lgl) remove missing data?}
}
\description{
\code{geom_timeline_label} return a \code{\link[ggplot2]{layer}}
    representing earthquakes annotations to be added after
    \code{\link{geom_timeline}}, i.e. a vertical line to each data
    point drown by \code{\link{geom_timeline}} with a text annotation
    (e.g. the location of the earthquake) attached to each earthquake
    line.
}
\details{
Aesthetics:
\code{geom_timeline_label} understands the following aesthetics
  (required in bold):
\itemize{
  \item \strong{x}: (Date) of earthquakes.
  \item \strong{label}: (chr) text annotation.
  \item y: (factr) stratification. If present multiple time lines
           will be plotted for each level of the factor
           (e.g. country).
  \item colour: of the points.
  \item n_max: annotation to drowm.
  \item size: of the points (if provided with \code{n_max} too,
        the \code{n_max} largest eartquakes will be annotated).
}
}
\examples{
\dontrun{
    library(dplyr)
    library(lubridate)
    library(ggplot2)

    library(devrcap)
        data(noaa)

    noaa \%>\%
        eq_clean_data() \%>\%
        filter(
            country \%in\% c("ITALY", "GREECE", "PORTUGAL"),
            year(date) >= 1900
        ) \%>\%
        ggplot(aes(
            x = date,
            y = country,
            size   = eq_primary,
            colour = log(total_deaths),
            label  = location_name
        )) +
        geom_timeline() +
        geom_timeline_label(n_max = 3)
}
}
\keyword{datasets}
