% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Graph the earthquake timeline for the US and China, with high magnitudes noted.}
\usage{
geom_timeline_label()
}
\description{
Graph the earthquake timeline for the US and China, with high magnitudes noted.
}
\examples{
geom_timeline_label()

}
