% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomTimelineLabel.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Label points on plot}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{x}{Date of earthquakes}

\item{size}{[optional] Size of earthquake}

\item{group}{[optional] Column indicating how the data could be grouped.}

\item{n_max}{[optional] number of top values to look at}

\item{label}{[optional] column to label the points.}
}
\value{
This function returns a graphic object
}
\description{
This method enchances the plot timeline function by labelling points.
By default, it will label the top n points.
}
\examples{
library(dplyr)
library(ggplot2)
raw_data<-readr::read_tsv(system.file("extdata", "signif.txt", package = "RCapstone"))
eq_clean_data(raw_data) \%>\% eq_location_clean() \%>\%
dplyr::filter(!is.na(EQ_PRIMARY), !is.na(DEATHS)) \%>\%
ggplot2::ggplot() +
 ggplot2::aes(
   x = DATE,
   size = EQ_PRIMARY,
   colour = DEATHS,
   label=LOCATION_NAME
 ) +
 geom_timeline() +
 geom_timeline_label()

}
