% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Mapping}

\item{data}{Data}

\item{stat}{Stat}

\item{position}{Position}

\item{na.rm}{na.rm}

\item{show.legend}{Show legend}

\item{inherit.aes}{Inherit aes}

\item{...}{other arguments}
}
\value{
Earthquake's labels
}
\description{
This function adds location labels
}
\examples{
\dontrun{
eq_location_clean(eq_clean_data(raw_data)) \%>\%
dplyr::filter(datetime >= "1960-01-01" & datetime <="2011-01-01" & COUNTRY == c("CHILE","USA"))\%>\%
ggplot() +
geom_timeline(aes(x = datetime, y = COUNTRY, size = EQ_MAG_ML, colour = DEATHS, fill = DEATHS)) +
geom_timeline_label(aes(x = datetime, y = COUNTRY, label = LOCATION_NAME, number = 3, max_aes = EQ_MAG_ML))
}

}
