% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Funcion for adding the Eartquakes's Location labels to an Earthquake's timeline}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, na.rm = TRUE,
  show.legend = NA, stat = "identity", position = "identity",
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{aesthetic mappings created by aes}

\item{data}{is the dataframe that contains the Earthquake's data}

\item{na.rm}{will hepls to remove the NA values from the data frame}

\item{show.legend}{layer's legend}

\item{stat}{The Layer's statistical transformation}

\item{position}{position adjustment functio}

\item{inherit.aes}{will indicate the default aesthetics overridng}

\item{...}{layer's other arguments}
}
\value{
the Earthquake's labels
}
\description{
Funcion for adding the Eartquakes's Location labels to an Earthquake's timeline
}
\examples{
\dontrun{
filename<-system.file("data","earthquakes_data.txt.zip",package="capstone")
eq_location_clean(eq_clean_data(eq_data_read(filename))) \%>\%
dplyr::filter(datetime >= "1980-01-01" & datetime <="2014-01-01" & COUNTRY == c("MEXICO","USA", "JORDAN"))\%>\%
ggplot() +
geom_timeline(aes(x = datetime, y = COUNTRY, size = EQ_MAG_ML, colour = DEATHS, fill = DEATHS)) +
geom_timeline_label(aes(x = datetime, y = COUNTRY, label = LOCATION_NAME, number = 3, max_aes = EQ_MAG_ML))
}

}
