% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomTimeline.R
\docType{data}
\name{GeomTIMELINE}
\alias{GeomTIMELINE}
\title{Plot Timeline - base packagae}
\format{An object of class \code{GeomTIMELINE} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\description{
Plot Timeline - base packagae
}
\keyword{datasets}
