% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Funcion for ploting an Earthquake's Location timeline building a GEOM Function from scratch}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\description{
Funcion for ploting an Earthquake's Location timeline building a GEOM Function from scratch
}
\keyword{datasets}
