% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Funcion for ploting an Earthquake's Location timeline building a GEOM Function from scratch
The GeomTimeLine will use a Dataframe compiled using the function eq_clean_data.
The GeomTimeLine function is a prototype function which will be used as foundation for our geom_timeline function.
The GeomTimeLine function will take advantage of the ggplot2's geom_point.
Using the Earthquakes' dates as X-axis main values, the Y-axis value will be not relevant while plotting a timeline horizontal bar
The geom_point's size and colour will be defined by the Earthquake's magnitude
The GeomTimeLine was build using the Function Prototype provided in the Course's Material 4.7.1 Building a New Geom}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeline
}
\description{
Funcion for ploting an Earthquake's Location timeline building a GEOM Function from scratch
The GeomTimeLine will use a Dataframe compiled using the function eq_clean_data.
The GeomTimeLine function is a prototype function which will be used as foundation for our geom_timeline function.
The GeomTimeLine function will take advantage of the ggplot2's geom_point.
Using the Earthquakes' dates as X-axis main values, the Y-axis value will be not relevant while plotting a timeline horizontal bar
The geom_point's size and colour will be defined by the Earthquake's magnitude
The GeomTimeLine was build using the Function Prototype provided in the Course's Material 4.7.1 Building a New Geom
}
\keyword{datasets}
