% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomTimelineLabel.R
\docType{data}
\name{GeomTIMELINELABEL}
\alias{GeomTIMELINELABEL}
\title{Label points on plot - Base package}
\format{An object of class \code{GeomTIMELINELABEL} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\description{
Label points on plot - Base package
}
\keyword{datasets}
