% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-eq.R
\name{theme_eq}
\alias{theme_eq}
\title{Minimal theme for timeline with earthquakes}
\usage{
theme_eq(...)
}
\arguments{
\item{...}{All options piped to \code{ggplot2::theme_minimal()}}
}
\description{
Minimal theme for timeline with earthquakes
}
\examples{
\dontrun{
library(ggplot2)
ggplot(noaa_data, aes(x = date, size = EQ_PRIMARY, n_max = 5)) +
  geom_timeline() +
  theme_eq()
}
}
